Finished trait Merge.
Made Downloads::add_directory() take a &Directory instead of Directory.
FileBases are cloned to add them to downloads.
This is because I don't want to consume the Directory,
so I can consume the Directory in the merge().
The merge() consumes the whole Directory,
while add_directory() only copies FileBases, so this makes sense.

Merge assumes we're merging a diff, so it's pretty simple,
although impl Merge for Directory is complicated b/c of the borrow checker, but I got it to work.